# backend/swagger.yaml
openapi: 3.0.3
info:
  title: API del Restaurante
  version: "1.0.0"
servers:
  - url: "http://localhost:3000"
paths:
...
  /staff:
    get:
      summary: Listar empleados
      responses:
        '200':
          description: OK
    post:
      summary: Crear empleado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                status:
                  type: string
                role:
                  type: string
      responses:
        '201':
          description: Creado
  /staff/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Obtener empleado
      responses:
        '200':
          description: OK
    put:
      summary: Actualizar empleado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                status:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: Actualizado
    delete:
      summary: Eliminar empleado
      responses:
        '200':
          description: Eliminado
...